Refactored GUI parsing page and gathered all parsing related functions in parse_data module
