+++ bmgui/main_page.py
@@ -7,6 +7,7 @@ import threading
 import tkinter as tk
 from functools import partial
 from pathlib import Path
+from tkinter.messagebox import showerror
 
 # 3rd party imports
 from screeninfo import get_monitors
